kafka-producer/requirements.txt
kafka-python
websocket-client
